requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
python-decouple
psycopg2-binary==2.9.9

# File handling and storage
django-storages==1.14.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23

# Text processing and embeddings (OpenAI + LangChain)
langchain
langchain-openai
langchain-community
openai
tiktoken

# Pinecone integration
pinecone
langchain-pinecone

# Additional utilities
python-magic
chardet